@@ -14,19 +14,17 @@
     "@emotion/styled": "11.14.0",
     "@mdi/js": "7.4.47",
     "@mdi/react": "1.6.1",
-    "@mui/material": "5.4.2",
-    "@mui/styled-engine": "5.4.2",
-    "@mui/styles": "5.4.2",
-    "@mui/system": "5.4.2",
+    "@mui/material": "5.17.1",
+    "@mui/system": "5.17.1",
     "i18next": "25.2.1",
     "lodash": "4.17.21",
     "mobx": "6.13.7",
     "mobx-react": "9.2.0",
-    "notistack": "2.0.8",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
+    "notistack": "3.0.2",
+    "react": "18.3.1",
+    "react-dom": "18.3.1",
     "react-i18next": "14.1.3",
-    "react-router-dom": "5.2.0"
+    "react-router-dom": "6.30.1"
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3.3.1",
@@ -36,9 +34,8 @@
     "@testing-library/user-event": "14.4.3",
     "@types/jest": "29.5.12",
     "@types/lodash": "4.17.17",
-    "@types/react": "17.0.20",
-    "@types/react-dom": "17.0.9",
-    "@types/react-router-dom": "^5.3.3",
+    "@types/react": "18.3.23",
+    "@types/react-dom": "18.3.7",
     "@typescript-eslint/eslint-plugin": "8.33.0",
     "@typescript-eslint/parser": "8.33.0",
     "@vitejs/plugin-react": "^4.2.1",
@@ -68,7 +65,8 @@
     "lint": "eslint src --ext .js,.jsx,.ts,.tsx",
     "format": "prettier --write .",
     "prepare": "husky",
-    "test": "jest --passWithNoTests"
+    "test": "jest --passWithNoTests",
+    "clean:install": "rm -rf node_modules pnpm-lock.yaml && pnpm install"
   },
   "lint-staged": {
     "**/*.{js,ts,jsx,tsx}": [
