@@ -1,12 +1,12 @@
-import ReactDOM from 'react-dom';
-
+import { createRoot } from 'react-dom/client';
 import App from './App';
 import './styles.css';
 
-const rootElement = document.getElementById('root');
+const container = document.getElementById('root');
 
-const render = (App: any) => {
-  ReactDOM.render(<App />, rootElement);
-};
-
-render(App);
+if (container) {
+  const root = createRoot(container);
+  root.render(<App />);
+} else {
+  console.error('Root element not found');
+}
